chore: migrate unpublish script -> deprecation script
